requirements.txt
Flask==3.0.0
python-docx==1.1.0
gunicorn==21.2.0

12313
